+++ src/Storage/Commands.Storage/Common/Cmdlet/NewAzureStorageContext.cs
@@ -16,6 +16,7 @@ using Microsoft.Azure.Commands.Common.Authentication;
 using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
 using Microsoft.Azure.ServiceManagemenet.Common;
 using Microsoft.WindowsAzure.Commands.Common;
+using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 using Microsoft.WindowsAzure.Commands.Common.Storage;
 using Microsoft.WindowsAzure.Commands.Utilities.Common;
 using Microsoft.WindowsAzure.Storage;
@@ -33,6 +34,7 @@ namespace Microsoft.WindowsAzure.Commands.Storage.Common.Cmdlet
     /// <summary>
     /// New storage context
     /// </summary>
+    [GenericBreakingChange("Default parameter set has changed from 'AccountNameAndKey' to 'OAuthAccount'")]
     [Cmdlet(VerbsCommon.New, StorageNouns.StorageContext, DefaultParameterSetName = OAuthParameterSet),
         OutputType(typeof(AzureStorageContext))]
     public class NewAzureStorageContext : AzureDataCmdlet